+++ tests/test_cases/objective_c/method_scope/input/source.m
@@ -1,0 +1,10 @@
+@implementation Example
+
+- (void)update:(int)value
+{
+    int localValue = value;
+finished:
+    return;
+}
+
+@end