+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add native Rust based Zig parser that uses WASM tree-sitter grammar
 - WASM Grammar management commands
 Allow users to manage the downloading of WASM grammars not bundled with treetags
+- Add native Rust based Objective C parser that uses WASM tree-sitter grammar
 
 ### Changed
 - Tag Address is truncated to 96 chars for native parsers
@@ -18,6 +19,8 @@ Allow users to manage the downloading of WASM grammars not bundled with treetags
 ### Removed
 - Removed the ability to use treetags as a library crate
 - Ocaml grammar bundled with treetags.
+- Zig plugin removed from the source repository
+- Objective C plugin removed from the source repository
 
 - [0.18.0]